work on test for workouts service

+++ backend/workouts/src/services/program.ts
@@ -1,27 +1,28 @@
-import { TokenPayload } from "google-auth-library";
-import { InternalUser, InternalUserDoc } from "../models/internal-user.model";
+import { InternalUserDoc } from "../models/internal-user.model";
 import { Program, ProgramDoc, WorkoutAttrs } from "../models/program.model";
 
 class WorkoutProgramService {
-   
-    private async getProgramBy(param: string) {
-        return await Program.findOne({[param]: param});
-    }
-
-    private async deleteProgramById(id: string) {
-        const program = await Program.findById(id);
-        if(!program) throw Error('Program not found');
-
-        await Program.deleteOne({_id: id});
-    }
-
     private async addProgramToUser(program: ProgramDoc, user: InternalUserDoc) {
         user.programs?.push(program);
         await user.save();
     }
+    
+    async deleteProgramById(user: InternalUserDoc, id: string) {
+        const programs = await this.getPrograms(user);
+        if(!programs) throw Error('User has no programs not found');
+
+        const program = programs.find(program => program._id.toString() === id);
+        if(!program) throw Error('Program not found');
+        
+        await Program.deleteOne({_id: id});
+        user.programs = user.programs?.filter(program => program._id.toString() !== id);
+        await user.save();
+    }
 
     async createProgram(name: string, endDate: string, workouts: WorkoutAttrs[], user: InternalUserDoc) {
-        if(await this.getProgramBy(name)) throw Error('Program with this name already exists');
+        const userPrograms = await this.getPrograms(user);
+        const isExistByName = userPrograms?.find(program => program.name === name);
+        if(isExistByName) throw Error('Program with this name already exists');
 
         const program = await Program.build({
             name,
@@ -31,7 +32,6 @@ class WorkoutProgramService {
         });
 
         await program.save();
-
         await this.addProgramToUser(program, user);
 
         return program;
@@ -42,9 +42,13 @@ class WorkoutProgramService {
         return userPrograms.programs;
     }
 
-    async getWorkouts(user: InternalUserDoc) {
-        const userPrograms = await user.populate('programs');
-        const workouts = userPrograms.programs?.map(program =>  program.workouts);
+    async getWorkouts(user: InternalUserDoc, programId: string) {
+        const userPrograms = await this.getPrograms(user);
+        if(!userPrograms) throw Error('User has no programs');
+
+        const workouts = userPrograms
+            .filter(program => program._id.toString() === programId)
+            .map(program =>  program.workouts);
         return workouts?.flat();
     }
 }
